+++ src/voice.hpp
@@ -21,33 +21,33 @@ class Voice {
 public:
   enum class State { Off, On, Releasing }; // TODO (feat): remove Releasing?
 
-  f32 phase = 0;
-  f32 vol = 0;
+  f64 phase = 0;
+  f64 vol = 0;
 
   void strike(
-    const f32 frequency,
-    const f32 current_time,
-    const f32 interpolation_duration,
+    const f64 frequency,
+    const f64 current_time,
+    const f64 interpolation_duration,
     const Interpolation interpolation);
-  void release(const Synth& synth, const f32 current_time);
-  void proceed_phase(const Synth& synth, const f32 current_time, const f32 dt);
+  void release(const Synth& synth, const f64 current_time);
+  void proceed_phase(const Synth& synth, const f64 current_time, const f64 dt);
 
   const bool has_state(const State target_state) const;
-  const bool has_target_frequency(const f32 frequency) const;
+  const bool has_target_frequency(const f64 frequency) const;
 
-  const f32 sample(const Synth& synth, const f32 song_time) const;
+  const f64 sample(const Synth& synth, const f64 song_time) const;
 
 private:
-  const f32 get_frequency(const f32 song_time) const;
+  const f64 get_frequency(const f64 song_time) const;
 
   State state = State::Off;
-  f32 last_strike_time = 0;                 // TODO (feat): seconds type
-  f32 last_release_time = 0;                // TODO (feat): seconds type
-  f32 frequency_interpolation_duration = 0; // TODO (feat): seconds type
+  f64 last_strike_time = 0;                 // TODO (feat): seconds type
+  f64 last_release_time = 0;                // TODO (feat): seconds type
+  f64 frequency_interpolation_duration = 0; // TODO (feat): seconds type
   Interpolation frequency_interpolation = Interpolation::None;
-  f32 last_frequency = 0;   // TODO (feat): frequency type
-  f32 target_frequency = 0; // TODO (feat): frequency type
-  // InterpolatedValue<f32> frequency;
+  f64 last_frequency = 0;   // TODO (feat): frequency type
+  f64 target_frequency = 0; // TODO (feat): frequency type
+  // InterpolatedValue<f64> frequency;
 };
 
 #endif
